+++ TaiFighter/assets/scripts/Client/Events/ChangePerspectiveEvent.lua
@@ -1,0 +1,8 @@
+local ns = require("namespace")
+
+ns.changePerspectiveEvent = ns.class("changePerspectiveEvent")
+
+function ns.changePerspectiveEvent:initialize(sideview)
+	self.sideview = sideview
+	LOG("Firing ChangedPerspectiveEvent")
+end
