+++ TaiFighter/assets/scripts/Client/Systems/BulletSystem.lua
@@ -1,0 +1,20 @@
+local ns = require('namespace')
+
+local BulletSystem = ns.class("BulletSystem",ns.System)
+
+function BulletSystem:requires() return {"bullet"} end
+
+function BulletSystem:update(dt)
+	for _, entity in pairs(self.targets) do
+		local bulletInfo = entity:get("bullet")
+		local movement = vec3:new(bulletInfo.speed*dt,0,0)
+		entity.Transform:translate(movement)
+		bulletInfo.lifetime = bulletInfo.lifetime - 1
+		if(bulletInfo.lifetime <= 0) then
+			--delete entity
+			Manager:removeEntity(entity)
+		end
+	end
+end
+
+Manager:addSystem(BulletSystem())
